Add new commands for sorting and separating YAML files with CLI integration
