+++ src/commands/mod.rs
@@ -1,0 +1,3 @@
+pub mod sort;
+pub mod separate;
+pub mod legacy;
